+++ modules/yarpc/doc.go
@@ -46,7 +46,7 @@
 //   func main() {
 //     svc, err := service.WithModule(
 //       "example",
-//       yarpc.ThriftModule(yarpc.CreateThriftServiceFunc(NewMyServiceHandler)),
+//       yarpc.New(yarpc.CreateThriftServiceFunc(NewMyServiceHandler)),
 //       service.WithModuleRole("service"),
 //     ).Build()
 //