+++ service/doc.go
@@ -45,7 +45,7 @@
 //       // The list of module creators for this service, in this case
 //       // creates a Thrift RPC module called "keyvalue"
 //       "keyvalue",
-//       yarpc.ThriftModule(yarpc.CreateThriftServiceFunc(NewYarpcThriftHandler)),
+//       yarpc.New(yarpc.CreateThriftServiceFunc(NewYarpcThriftHandler)),
 //     ).Build()
 //
 //     if err != nil {
@@ -79,7 +79,7 @@
 //       kafka.Module("kakfa_topic1", []string{"worker"}),
 //     ).WithModule(
 //       "keyvalue",
-//       rpc.ThriftModule(rpc.CreateThriftServiceFunc(NewYarpcThriftHandler)),
+//       yarpc.New(rpc.CreateThriftServiceFunc(NewYarpcThriftHandler)),
 //       service.WithModuleRole("service"),
 //     ).Build()
 //